Increase ping timeout interval
The previous WiFi patch that restricted AP mode to only when there is no WiFi config causes issues with gateways with exisiting WiFi config.
Though the two are not directly related, the patch seems to have uncovered an existing problem.
In the current code, the WiFi ping timeout is 8s (4x2 tries). But practically, it takes 10-15s to re-connect. So, after 8s, we go to WAIT_FOR_NETWORK state where the WiFi module is restarted.
This causes a loop since everytime the pings timeout at 8s and restart the WiFi module.
By increasing the ping timeout to a higher value of 120s (60x2 tries), this issue is not observed since the WiFi connects after roughly after 15-20s.

This issue was not observed in gateways where the WiFi config was erased since the initial wait for connection already takes into account the extra time needed to establish a connection.

Two things to be considered in the future
- Select an appropriate value for the timeout based on testing instead an arbitirarily chosen value.
- Solve the actual problem of the WiFi driver (which is beyond the scope of the Application code)